+++ internal/tooling/templates/embed.go
@@ -1,0 +1,6 @@
+package templates
+
+import "embed"
+
+//go:embed templates/**
+var embeddedTemplates embed.FS