+++ internal/tooling/templates/templates/nuxt-ts/app.vue
@@ -1,0 +1,6 @@
+<template>
+  <div class="app">
+    <h1>AitiGo Nuxt 3 Template</h1>
+    <p>Start building your app here.</p>
+  </div>
+</template>